+++ ksqlDb.RestApi.Client/KSql/RestApi/Parsers/IdentifierUtil.cs
@@ -1,5 +1,9 @@
+using System.Reflection;
 using Antlr4.Runtime;
+using ksqlDB.RestApi.Client.Infrastructure.Extensions;
 using ksqlDB.RestApi.Client.KSql.RestApi.Enums;
+using ksqlDB.RestApi.Client.KSql.RestApi.Statements.Annotations;
+using static ksqlDB.RestApi.Client.KSql.RestApi.Enums.IdentifierEscaping;
 
 namespace ksqlDb.RestApi.Client.KSql.RestApi.Parsers
 {
@@ -37,10 +41,30 @@ namespace ksqlDb.RestApi.Client.KSql.RestApi.Parsers
     {
       return escaping switch
       {
-        IdentifierEscaping.Never => identifier,
-        IdentifierEscaping.Keywords when IsValid(identifier) && SystemColumns.IsValid(identifier) => identifier,
-        IdentifierEscaping.Keywords => string.Concat("`", identifier, "`"),
-        IdentifierEscaping.Always => string.Concat("`", identifier, "`"),
+        Never => identifier,
+        Keywords when IsValid(identifier) && SystemColumns.IsValid(identifier) => identifier,
+        Keywords => string.Concat("`", identifier, "`"),
+        Always => string.Concat("`", identifier, "`"),
+        _ => throw new ArgumentOutOfRangeException(nameof(escaping), escaping, "Non-exhaustive match.")
+      };
+    }
+
+    /// <summary>
+    /// Format the <c>identifier</c>, except when it is a <c>PseudoColumn</c>.
+    /// </summary>
+    /// <param name="memberInfo">the memberInfo with the identifier</param>
+    /// <param name="escaping">the format</param>
+    /// <returns>the identifier modified based on the provided <c>format</c></returns>
+    public static string Format(MemberInfo memberInfo, IdentifierEscaping escaping)
+    {
+      return escaping switch
+      {
+        Never => memberInfo.GetMemberName(),
+        Keywords when memberInfo.GetCustomAttribute<PseudoColumnAttribute>() != null => memberInfo.Name,
+        Keywords when IsValid(memberInfo.GetMemberName()) && SystemColumns.IsValid(memberInfo.GetMemberName()) => memberInfo.GetMemberName(),
+        Keywords => string.Concat("`", memberInfo.GetMemberName(), "`"),
+        Always when memberInfo.GetCustomAttribute<PseudoColumnAttribute>() != null => memberInfo.Name,
+        Always => string.Concat("`", memberInfo.GetMemberName(), "`"),
         _ => throw new ArgumentOutOfRangeException(nameof(escaping), escaping, "Non-exhaustive match.")
       };
     }