+++ ksqlDb.RestApi.Client/KSql/RestApi/Statements/Annotations/PseudoColumnAttribute.cs
@@ -1,0 +1,6 @@
+namespace ksqlDB.RestApi.Client.KSql.RestApi.Statements.Annotations;
+
+[AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
+internal sealed class PseudoColumnAttribute : Attribute
+{
+}